Parse PG manifest parameters; let Portal.Context bind from the values.
First, a simple modification to geni.rspec.pgmanifest to help that class
extract PG portal parameters.  Users would call
Manifest(xml="<manifest>").parameters to get a generator yielding the
parameters.

Second, there's modifications to geni.portal.Context to allow a user to
supply an "altParamSrc" named argument to Context.bindParameters().
They can supply a dict of params; a geni.rspec.pgmanifest.Manifest
object (from which we extract the parameters and convert to a dict); or
a string (which we try to parse as a PG manifest).