Make geni.portal.Conext a singleton, and make a gloabal for it
Rationale: there is no reason to import the geni.portal module unless
you are going to make a Context() object, so make a global one
(geni.portal.context) automatically for cleaner user code.

But, due to some other behavior we are working on (namely, making
the portal.Context auto-print on exit), it could be messy to have
more than one of these objects. So, we make this class a singleton.

The older method of explicitly building a Context object will still
work (it will just result in getting a reference to the existing
object).