Allow user to bind a request RSpec to a portal.Context
This way, we can have functions like printRequestRSpec() operate on the bound
request rather than having to always pass them explicitly.

At this time, only one request can be bound to a context, though the API is
explicitly designed to permit such a thing in the future if we can figure out
appropriate semantics.

Add a convenience function to make a new request and bind it.