A bridged link is syntactic sugar used to create two links separated by an Emulab delay (bridge) node. The BridgedLink class will create the following topology:
	      left-link          right-link
	node1 =========== bridge ============ node2

The bridge is a special node type (sliver_type="delay") that tells the
CM to insert an Emulab delay node instead of a plain (router) node. A
delay node is a transparent Ethernet bridge between the left and right
segments above, but on which the traffic can be shaped wrt. bandwidth,
latency, and loss. For example:

        # Create the bridged link between the two nodes.
        link = request.BridgedLink("link")
        # Add two interfaces
        link.addInterface(iface1)
        link.addInterface(iface2)
        # Give the link (bridge) some shaping parameters.
        link.bandwidth = 10000
        link.latency   = 15
        link.plr       = 0.01