Ensure that any generated errors are reported before printing rspec.
Prior to this commit, we were not printing generated errors if
the user did not either call reportError(...,immediate=True) or
portal.context.verifyParameters().  Now printRequestRspec()
calls verifyParameters() to be safe.

There is a slight chance this could break old profiles, because those
could possibly have been generating errors that were not being reported.
But now they can find out that their code was broken!  That's an upside,
right?